feat:delete blog if token is correct

+++ controllers/blogs.js
@@ -44,9 +44,25 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) {
+    return response.status(204).end();
+  }
+
+  if (!(blog.user.toString() === decodedToken.id)) {
+    return response
+      .status(401)
+      .json({ error: "You do not have the permission to delete this blog." });
+  }
   await Blog.findByIdAndDelete(request.params.id);
 
-  response.status(204).end();
+  return response.status(204).end();
 });
 
 blogsRouter.put("/:id", async (request, response) => {
